Fix tx bus problem.
All targets share the same tx-rx lines. The RX lines float but, by default, the TX line has an internal pull-up. Whenever more than 2 targets were attached the logic-voltage of the tx line would only drop by 1/3rd (or 1/4th-1/5th, etc for the length of the bus) instead of going from VCC to 0. To fix this during the transmit I move the tx function to the tx line and when I'm done I move it to a spare IO and let the original tx pin float.