Switch ir pin back to PWM.
So I originally used a scope and reverse engineered the IR protocol noting a 40khz carrier freq. Months later, when I was debugging the code, I didn't have a scope and used a logic analyzer instead, but didn't see the carrier. After many hours of debug I decided to order new scope probes (since the old probes were lost in a house fire) and debug the protocol again. I found again, there is a carrier, and added it back in. Now it works and all is right in the world.